requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.10.1
pyOpenSSL==23.3.0
